Move keymap defaults to a separate file. Not exactly a clean separation though.
