+++ Core/Core.vcxproj
@@ -541,6 +541,7 @@
     <ClCompile Include="HW\Camera.cpp" />
     <ClCompile Include="Instance.cpp" />
     <ClCompile Include="KeyMap.cpp" />
+    <ClCompile Include="KeyMapDefaults.cpp" />
     <ClCompile Include="MemFault.cpp" />
     <ClCompile Include="MIPS\fake\FakeJit.cpp" />
     <ClCompile Include="MIPS\IR\IRAsm.cpp" />
@@ -1098,6 +1099,7 @@
     <ClInclude Include="HW\Camera.h" />
     <ClInclude Include="Instance.h" />
     <ClInclude Include="KeyMap.h" />
+    <ClInclude Include="KeyMapDefaults.h" />
     <ClInclude Include="MemFault.h" />
     <ClInclude Include="MIPS\fake\FakeJit.h" />
     <ClInclude Include="MIPS\IR\IRFrontend.h" />
