+++ Core/KeyMap.h
@@ -65,16 +65,6 @@ enum {
 	VIRTKEY_COUNT = VIRTKEY_LAST - VIRTKEY_FIRST
 };
 
-enum DefaultMaps {
-	DEFAULT_MAPPING_KEYBOARD,
-	DEFAULT_MAPPING_PAD,
-	DEFAULT_MAPPING_X360,
-	DEFAULT_MAPPING_SHIELD,
-	DEFAULT_MAPPING_OUYA,
-	DEFAULT_MAPPING_XPERIA_PLAY,
-	DEFAULT_MAPPING_MOQI_I7S,
-};
-
 const float AXIS_BIND_THRESHOLD = 0.75f;
 const float AXIS_BIND_THRESHOLD_MOUSE = 0.01f;
 
@@ -106,6 +96,7 @@ class IniFile;
 
 namespace KeyMap {
 	extern KeyMapping g_controllerMap;
+	extern std::set<int> g_seenDeviceIds;
 	extern int g_controllerMapGeneration;
 
 	// Key & Button names
