+++ Core/KeyMapDefaults.h
@@ -1,0 +1,13 @@
+#pragma once
+
+enum DefaultMaps {
+	DEFAULT_MAPPING_KEYBOARD,
+	DEFAULT_MAPPING_PAD,
+	DEFAULT_MAPPING_X360,
+	DEFAULT_MAPPING_SHIELD,
+	DEFAULT_MAPPING_OUYA,
+	DEFAULT_MAPPING_XPERIA_PLAY,
+	DEFAULT_MAPPING_MOQI_I7S,
+};
+
+void SetDefaultKeyMap(DefaultMaps dmap, bool replace);
