+++ android/jni/Android.mk
@@ -387,6 +387,7 @@ EXEC_AND_LIB_FILES := \
   $(SRC)/Core/HDRemaster.cpp \
   $(SRC)/Core/Instance.cpp \
   $(SRC)/Core/KeyMap.cpp \
+  $(SRC)/Core/KeyMapDefaults.cpp \
   $(SRC)/Core/Host.cpp \
   $(SRC)/Core/Loaders.cpp \
   $(SRC)/Core/PSPLoaders.cpp \
